+++ bnp_na_lib/__init__.py
@@ -1,2 +1,2 @@
-"""Helper modules for bnp_na V13.9."""
-__version__ = "V13.9"
+"""Helper modules for bnp_na V13.10."""
+__version__ = "V13.10"